+++ packages/experimental/src/components/ActionBar/ActionBar.stories.js
@@ -11,14 +11,16 @@ import { action } from '@storybook/addon-actions';
 import { ActionBar, ActionBarItem } from '.';
 import { Lightning16, Bee24 } from '@carbon/icons-react';
 
-import { pkgPrefix } from '../../global/js/settings';
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'ActionBar');
 
-const blockClass = `${pkgPrefix}-action-bar`;
+const blockClass = `${pkg.prefix}-action-bar`;
 
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 
 export default {
-  title: 'Experimental/ActionBar',
+  title: `${storybookPrefix}/PageHeader/ActionBarItem`,
   component: ActionBar,
   argTypes: {
     containerWidth: {