+++ packages/experimental/src/components/BreadcrumbWithOverflow/BreadcrumbWithOverflow.stories.js
@@ -13,11 +13,14 @@ import { BreadcrumbWithOverflow } from '.';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import { BreadcrumbItem } from 'carbon-components-react';
 
-import { pkgPrefix } from '../../global/js/settings';
-const blockClass = `${pkgPrefix}-breadcrumb-with-overflow`;
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'BreadcrumbWithOverflow');
+
+const blockClass = `${pkg.prefix}-breadcrumb-with-overflow`;
 
 export default {
-  title: 'Experimental/BreadcrumbWithOverflow',
+  title: `${storybookPrefix}/BreadcrumbWithOverflow`,
   component: BreadcrumbWithOverflow,
   argTypes: {
     containerWidth: {