+++ packages/experimental/src/components/Card/Card.stories.js
@@ -6,13 +6,18 @@
 //
 
 import React from 'react';
+import cx from 'classnames';
 import { Card } from '.';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import mdx from './Card.mdx';
 import { ArrowRight24, Cloud32 } from '@carbon/icons-react';
+import { AspectRatio } from 'carbon-components-react';
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'Card');
 
 export default {
-  title: 'Experimental/Card',
+  title: `${storybookPrefix}/Card`,
   component: Card,
   parameters: {
     styles,
@@ -20,6 +25,29 @@ export default {
       page: mdx,
     },
   },
+  argTypes: {
+    columnSize: {
+      defaultValue: 'lg-4',
+      control: {
+        type: 'select',
+        options: ['sm-4', 'md-4', 'lg-4', 'max-4'],
+      },
+    },
+    mediaRatio: {
+      defaultValue: '1x1',
+      control: {
+        type: 'select',
+        options: ['16x9', '9x16', '2x1', '1x2', '4x3', '3x4', '1x1'],
+      },
+    },
+  },
+  decorators: [
+    (Story) => (
+      <div className="bx--grid card-story">
+        <Story />
+      </div>
+    ),
+  ],
 };
 
 const defaultProps = {
@@ -32,23 +60,34 @@ const defaultProps = {
       action on the card.
     </p>
   ),
-  actionButtonText: 'Primary button',
+  primaryButtonText: 'Primary',
+  cols: 4,
 };
 
 const Template = (opts) => {
-  const { children, ...args } = opts;
+  const { children, columnSize, ...args } = opts;
+  const colClasses = cx(`bx--col-${columnSize}`);
   return (
-    <div className="card-demo-container">
-      <Card {...args}>{children}</Card>
+    <div className="bx--row">
+      <div className={colClasses}>
+        <Card {...args}>{children}</Card>
+      </div>
     </div>
   );
 };
 
-const TemplateWide = (opts) => {
-  const { children, ...args } = opts;
+const MediaTemplate = (opts) => {
+  const { children, columnSize, mediaRatio, ...args } = opts;
+  const colClasses = cx(`bx--col-${columnSize}`);
   return (
-    <div className="card-demo-container--wide">
-      <Card {...args}>{children}</Card>
+    <div className="bx--row">
+      <div className={colClasses}>
+        <Card
+          media={<AspectRatio ratio={mediaRatio}>{mediaRatio}</AspectRatio>}
+          {...args}>
+          {children}
+        </Card>
+      </div>
     </div>
   );
 };
@@ -56,21 +95,38 @@ const TemplateWide = (opts) => {
 export const Default = Template.bind({});
 Default.args = {
   ...defaultProps,
-  media: <img src="https://via.placeholder.com/300x200/000/fff" alt="img" />,
 };
 
-export const MediaLeft = TemplateWide.bind({});
+export const LabelOnly = Template.bind({});
+LabelOnly.args = {
+  ...defaultProps,
+  title: '',
+};
+
+export const WithCaption = Template.bind({});
+WithCaption.args = {
+  ...defaultProps,
+  caption: 'Description or long caption',
+  label: '',
+};
+
+export const WithMedia = MediaTemplate.bind({});
+WithMedia.args = {
+  ...defaultProps,
+};
+
+export const MediaLeft = MediaTemplate.bind({});
 MediaLeft.args = {
   ...defaultProps,
   mediaPosition: 'left',
-  media: <img src="https://via.placeholder.com/300x225/000/fff" alt="img" />,
+  columnSize: 'md-4',
 };
 
 export const WithActionIcon = Template.bind({});
 WithActionIcon.args = {
   ...defaultProps,
   actionIcon: ArrowRight24,
-  actionButtonText: '',
+  primaryButtonText: '',
 };
 
 export const WithPictogram = Template.bind({});
@@ -79,16 +135,17 @@ WithPictogram.args = {
   pictogram: Cloud32,
 };
 
+export const WithSecondaryAction = Template.bind({});
+WithSecondaryAction.args = {
+  ...defaultProps,
+  secondaryButtonText: 'Secondary',
+  secondaryButtonKind: 'ghost',
+  columnSize: 'md-4',
+};
+
 export const ClickableCardWithOnclick = Template.bind({});
 ClickableCardWithOnclick.args = {
   ...defaultProps,
   onClick: () => {},
-  actionButtonText: '',
-};
-
-export const ClickableCardWithLink = Template.bind({});
-ClickableCardWithLink.args = {
-  ...defaultProps,
-  href: '/',
-  actionButtonText: '',
+  primaryButtonText: '',
 };