+++ packages/experimental/src/components/ExportModal/ExportModal.stories.js
@@ -10,9 +10,12 @@ import { Button } from 'carbon-components-react';
 import { ExportModal } from '.';
 import styles from './_storybook-styles.scss'; // import index in case more files are added later.
 import mdx from './ExportModal.mdx';
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'ExportModal');
 
 export default {
-  title: 'Experimental/ExportModal',
+  title: `${storybookPrefix}/ExportModal`,
   component: ExportModal,
   parameters: {
     styles,