+++ packages/experimental/src/components/ImportModal/ImportModal.js
@@ -7,7 +7,7 @@ import {
   Button,
 } from 'carbon-components-react';
 import PropTypes from 'prop-types';
-import { pkgPrefix } from '../../global/js/settings';
+import { pkg } from '../../settings';
 import uuidv4 from '../../global/js/utils/uuidv4';
 
 export const ImportModal = ({
@@ -141,17 +141,17 @@ export const ImportModal = ({
       primaryButtonDisabled={primaryButtonDisabled}
       onRequestSubmit={onSubmitHandler}
       onRequestClose={onRequestClose}
-      className={`${pkgPrefix}--import-modal`}>
-      <p className={`${pkgPrefix}--import-modal-body`}>{modalBody}</p>
-      <p className={`${pkgPrefix}--import-modal-label`}>{fileDropHeader}</p>
+      className={`${pkg.prefix}--import-modal`}>
+      <p className={`${pkg.prefix}--import-modal-body`}>{modalBody}</p>
+      <p className={`${pkg.prefix}--import-modal-label`}>{fileDropHeader}</p>
       <FileUploaderDropContainer
         accept={validFileTypes}
         labelText={fileDropLabel}
         onAddFiles={onAddFile}
         disabled={files.length}
       />
-      <p className={`${pkgPrefix}--import-modal-label`}>{inputHeader}</p>
-      <div className={`${pkgPrefix}--input-group`}>
+      <p className={`${pkg.prefix}--import-modal-label`}>{inputHeader}</p>
+      <div className={`${pkg.prefix}--input-group`}>
         <TextInput
           id={inputId}
           labelText={inputLabel}
@@ -162,7 +162,7 @@ export const ImportModal = ({
         />
         <Button
           onClick={fetchFile}
-          className={`${pkgPrefix}--import-button`}
+          className={`${pkg.prefix}--import-button`}
           size="sm"
           disabled={importButtonDisabled}>
           {inputButtonText}
@@ -170,7 +170,7 @@ export const ImportModal = ({
       </div>
       <div className="bx--file-container" style={{ width: '100%' }}>
         {fileUploaded && (
-          <p className={`${pkgPrefix}--import-modal-helper-text`}>
+          <p className={`${pkg.prefix}--import-modal-helper-text`}>
             {fileUploadLabel}
           </p>
         )}