+++ packages/experimental/src/components/Notifications/Notification.stories.js
@@ -17,13 +17,17 @@ import {
 import { User20, Notification20 } from '@carbon/icons-react';
 import { white } from '@carbon/colors';
 import styles from './_storybook-styles.scss';
-
 import { Notifications } from '.';
+
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'Notifications');
+
 import mdx from './Notifications.mdx';
 import data from './Notifications_data';
 
 export default {
-  title: 'Experimental/Notifications',
+  title: `${storybookPrefix}/Notifications`,
   component: Notifications,
   parameters: {
     styles,