+++ packages/experimental/src/components/Tearsheet/TearsheetNarrow.stories.js
@@ -9,7 +9,9 @@ import React, { useState } from 'react';
 
 import { action } from '@storybook/addon-actions';
 
-import { pkgPrefix } from '../../global/js/settings';
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'TearsheetNarrow');
 
 import { Button } from 'carbon-components-react';
 
@@ -21,7 +23,7 @@ import styles from './_storybook-styles.scss';
 import mdx from './Tearsheet.mdx';
 
 export default {
-  title: 'Experimental/Tearsheets/TearsheetNarrow',
+  title: `${storybookPrefix}/Tearsheets/TearsheetNarrow`,
   component: TearsheetNarrow,
   subcomponents: {
     Tearsheet,
@@ -202,7 +204,7 @@ const Template = ({ buttonSet, ...args }) => {
 
   return (
     <>
-      <style>{`.${pkgPrefix}-tearsheet { opacity: 0 }`};</style>
+      <style>{`.${pkg.prefix}-tearsheet { opacity: 0 }`};</style>
       <Button onClick={() => setOpen(true)}>Open Tearsheet</Button>
       <TearsheetNarrow
         className={className}