+++ packages/experimental/src/components/Tearsheet/TearsheetShell.stories.js
@@ -10,11 +10,14 @@ import React from 'react';
 import { TearsheetShell } from './TearsheetShell';
 
 import styles from './_storybook-styles.scss';
+import { pkg } from '../../settings';
+import { getStorybookPrefix } from '../../../config';
+const storybookPrefix = getStorybookPrefix(pkg, 'TearsheetShell');
 
 import mdx from './TearsheetShell.mdx';
 
 export default {
-  title: 'Experimental/Tearsheets/TearsheetShell',
+  title: `${storybookPrefix}/Tearsheets/TearsheetShell`,
   component: TearsheetShell,
   parameters: { controls: { expanded: true }, styles, docs: { page: mdx } },
 };