+++ packages/experimental/src/global/js/settings.js
@@ -1,7 +1,0 @@
-import featureFlags from '../../generated/feature-flags/feature-flags';
-import { settings } from 'carbon-components';
-
-const pkgPrefix = 'exp';
-const carbonPrefix = settings.prefix;
-
-export { pkgPrefix, carbonPrefix, featureFlags };